+++ src/page-objects/AdministrationPages.ts
@@ -106,8 +106,8 @@ export const AdminPageObjects = {
 
 export const test = base.extend<FixtureTypes>({
 
-    AdminProductDetail: async ({ AdminPage }, use) => {
-        await use(new ProductDetail(AdminPage));
+    AdminProductDetail: async ({ AdminPage, InstanceMeta }, use) => {
+        await use(new ProductDetail(AdminPage, InstanceMeta));
     },
 
     AdminOrderDetail: async ({ AdminPage }, use) => {
@@ -126,12 +126,12 @@ export const test = base.extend<FixtureTypes>({
         await use(new CustomerGroupListing(AdminPage));
     },
 
-    AdminCustomerGroupCreate: async ({ AdminPage }, use) => {
-        await use(new CustomerGroupCreate(AdminPage));
+    AdminCustomerGroupCreate: async ({ AdminPage, InstanceMeta }, use) => {
+        await use(new CustomerGroupCreate(AdminPage, InstanceMeta));
     },
 
-    AdminCustomerGroupDetail: async ({ AdminPage }, use) => {
-        await use(new CustomerGroupDetail(AdminPage));
+    AdminCustomerGroupDetail: async ({ AdminPage, InstanceMeta }, use) => {
+        await use(new CustomerGroupDetail(AdminPage, InstanceMeta));
     },
 
     AdminFirstRunWizard: async ({ AdminPage, InstanceMeta }, use) => {
@@ -174,8 +174,8 @@ export const test = base.extend<FixtureTypes>({
         await use(new Categories(AdminPage));
     },
 
-    AdminCategoryDetail: async ({ AdminPage }, use) => {
-        await use(new CategoryDetail(AdminPage));
+    AdminCategoryDetail: async ({ AdminPage, InstanceMeta }, use) => {
+        await use(new CategoryDetail(AdminPage, InstanceMeta));
     },
 
     AdminLandingPageDetail: async ({ AdminPage }, use) => {