+++ src/page-objects/administration/ManufacturerDetail.ts
@@ -1,6 +1,7 @@
 import type { Page, Locator } from '@playwright/test';
 import { ManufacturerCreate } from './ManufacturerCreate';
 import { HelperFixtureTypes } from 'src/fixtures/HelperFixtures';
+import { satisfies } from 'compare-versions';
 
 export class ManufacturerDetail extends ManufacturerCreate {
     public readonly customFieldCard: Locator;
@@ -10,9 +11,31 @@ export class ManufacturerDetail extends ManufacturerCreate {
     constructor(public readonly page: Page, public readonly instanceMeta: HelperFixtureTypes['InstanceMeta']) {
         super(page, instanceMeta);
 
-        this.customFieldCard = page.locator('.sw-card').getByText('Custom fields');
+        if (satisfies(instanceMeta.version, '<6.7')) {
+            this.customFieldCard = page.locator('.sw-card').getByText('Custom fields');
+        } else {
+            this.customFieldCard = page.locator('.mt-card').getByText('Custom fields');
+        }
+
         this.customFieldSetTabs = this.customFieldCard.locator('.sw-tabs-item');
         this.customFieldSetTabCustomContent = this.customFieldCard.locator('.sw-tabs__custom-content');
+    }
+
+    async getCustomFieldSetCardContentByName(customFieldSetName: string): Promise<Record<string, Locator>> {
+        let customFieldCard: Locator;
+        if (satisfies(this.instanceMeta.version, '<6.7')) {
+            customFieldCard = this.page.locator('.mt-card').filter({ hasText: 'Custom fields' });
+        } else {
+            customFieldCard = this.page.locator('.mt-card').filter({ hasText: 'Custom fields' });
+        }
+
+        const customFieldSetTab = customFieldCard.getByText(customFieldSetName);
+        const customFieldSetTabCustomContent = customFieldCard.locator(`.sw-custom-field-set-renderer-tab-content__${customFieldSetName}`);
+
+        return {
+            customFieldSetTab: customFieldSetTab,
+            customFieldSetTabCustomContent: customFieldSetTabCustomContent,
+        }
     }
 
     url(manufacturerUuid?: string) {