+++ src/types/ShopwareTypes.ts
@@ -35,6 +35,19 @@ export type CustomerAddress = components['schemas']['CustomerAddress'] & {
     id: string,
 }
 
+export interface Address {
+    salutation: string,
+    firstName: string,
+    lastName: string,
+    company: string,
+    department: string,
+    street: string,
+    city: string,
+    zipCode: string,
+    country: string,
+    state: string,
+}
+
 export type Salutation = components['schemas']['Salutation'] & {
     id: string,
 }
@@ -46,6 +59,10 @@ export interface Price {
     currencyId: string;
 }
 
+export interface VariantListingConfig {
+    displayParent: boolean;
+}
+
 export interface ProductPrice {
     productId?: string;
     ruleId: string;
@@ -54,7 +71,7 @@ export interface ProductPrice {
     quantityEnd: number | null;
 }
 
-export type Product = Omit<components['schemas']['Product'], 'price' | 'prices' | 'options' | 'tags' | 'visibilities' > & {
+export type Product = Omit<components['schemas']['Product'], 'price' | 'prices' | 'options' | 'tags' | 'visibilities' | 'variantListingConfig' > & {
     id: string,
     price: Price[],
     prices?: ProductPrice[],
@@ -64,6 +81,7 @@ export type Product = Omit<components['schemas']['Product'], 'price' | 'prices' 
     options?: Record<string, string>[],
     tags?: Record<string, string>[],
     visibilities?: Record<string, unknown>[],
+    variantListingConfig?: VariantListingConfig,
 }
 
 export type OrderDelivery = Omit<components['schemas']['OrderDelivery'], 'shippingOrderAddress' | 'shippingCosts'> & {
@@ -106,13 +124,21 @@ export type Currency = components['schemas']['Currency'] & {
     id: string,
 }
 
-export type Country = components['schemas']['Country'] & {
+export type Country = Omit<components['schemas']['Country'], 'states'> & {
     id: string,
+    states: [{
+        name: string,
+        shortCode: string,
+    }],
 }
 
 export type SystemConfig = components['schemas']['SystemConfig'] & {
     id: string,
 }
+
+export type ProductCrossSelling = components['schemas']['ProductCrossSelling'] & {
+    id: string,
+};
 
 export interface CalculatedTaxes {
     tax: number,